Add fsck.mode=skip to kernel command line (#26)
By default the autoinstaller performs a checksum verification after
booting before continuing with the installation, which can take a couple
of minutes, in particular when using a virtual CDROM drive as boot
device. Skip this verification since error rates are usually small
enough, in particular when no physical devices are involved. The
packages that are installed onto the target system are independently
verified later anyway.

Signed-off-by: Dr. Jens Harbott <harbott@osism.tech>